[ refactor ] 예약 생성 요청 DTO 유효성 검사 어노테이션 변경 ( #3 )
