+++ src/main/java/com/qring/reservation/presentation/v1/req/PostReservationReqDTOV1.java
@@ -1,9 +1,7 @@
 package com.qring.reservation.presentation.v1.req;
 
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.Max;
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -18,7 +16,7 @@ public class PostReservationReqDTOV1 {
     @Getter
     public static class Reservation {
 
-        @NotNull(message = "인원 수를 입력해주세요.")
+        @Positive(message = "인원 수를 입력해주세요.")
         @Min(value = 1, message = "예약 인원은 최소 1인입니다.")
         @Max(value = 6, message = "예약 인원은 최대 6인입니다.")
         private int headCount;
